Renamed package names to meet Java naming conventions
