+++ src/main/java/org/mitre/synthea/export/CSVExporter.java
@@ -54,8 +54,8 @@ import org.mitre.synthea.world.concepts.HealthRecord.Medication;
 import org.mitre.synthea.world.concepts.HealthRecord.Observation;
 import org.mitre.synthea.world.concepts.HealthRecord.Procedure;
 import org.mitre.synthea.world.concepts.HealthRecord.Supply;
-import org.mitre.synthea.world.concepts.health_insurance.Claim;
-import org.mitre.synthea.world.concepts.health_insurance.CoverageRecord.PlanRecord;
+import org.mitre.synthea.world.concepts.healthinsurance.Claim;
+import org.mitre.synthea.world.concepts.healthinsurance.CoverageRecord.PlanRecord;
 
 
 /**
