+++ src/main/java/org/mitre/synthea/world/agents/Person.java
@@ -32,8 +32,8 @@ import org.mitre.synthea.world.concepts.HealthRecord;
 import org.mitre.synthea.world.concepts.HealthRecord.Code;
 import org.mitre.synthea.world.concepts.HealthRecord.Encounter;
 import org.mitre.synthea.world.concepts.HealthRecord.EncounterType;
-import org.mitre.synthea.world.concepts.health_insurance.CoverageRecord;
-import org.mitre.synthea.world.concepts.health_insurance.CoverageRecord.PlanRecord;
+import org.mitre.synthea.world.concepts.healthinsurance.CoverageRecord;
+import org.mitre.synthea.world.concepts.healthinsurance.CoverageRecord.PlanRecord;
 import org.mitre.synthea.world.concepts.VitalSign;
 import org.mitre.synthea.world.geography.quadtree.QuadTreeElement;
