+++ src/main/java/org/mitre/synthea/world/agents/behaviors/payeradjustment/IPayerAdjustment.java
@@ -1,7 +1,7 @@
-package org.mitre.synthea.world.agents.behaviors.payer_adjustment;
+package org.mitre.synthea.world.agents.behaviors.payeradjustment;
 
 import org.mitre.synthea.world.agents.Person;
-import org.mitre.synthea.world.concepts.health_insurance.Claim.ClaimEntry;
+import org.mitre.synthea.world.concepts.healthinsurance.Claim.ClaimEntry;
 
 /**
  * Interface to abstract various claim adjustment strategies.
