+++ src/main/java/org/mitre/synthea/world/agents/behaviors/providerfinder/ProviderFinderRandom.java
@@ -1,4 +1,4 @@
-package org.mitre.synthea.world.agents.behaviors.provider_finder;
+package org.mitre.synthea.world.agents.behaviors.providerfinder;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -7,25 +7,16 @@ import org.mitre.synthea.world.agents.Person;
 import org.mitre.synthea.world.agents.Provider;
 import org.mitre.synthea.world.concepts.HealthRecord.EncounterType;
 
-public class ProviderFinderQuality implements IProviderFinder {
-
-  private IProviderFinder nearest = new ProviderFinderNearest();
+public class ProviderFinderRandom implements IProviderFinder {
 
   @Override
   public Provider find(List<Provider> providers, Person person, EncounterType service, long time) {
-    List<Provider> options = new ArrayList<>();
-    double bestQuality = Double.NEGATIVE_INFINITY;
+    List<Provider> options = new ArrayList<Provider>();
 
     for (Provider provider : providers) {
       if (provider.accepts(person, time)
           && (provider.hasService(service) || service == null)) {
-        if (provider.quality > bestQuality) {
-          options.clear();
-          options.add(provider);
-          bestQuality = provider.quality;
-        } else if (provider.quality == bestQuality) {
-          options.add(provider);
-        }
+        options.add(provider);
       }
     }
 
@@ -34,7 +25,8 @@ public class ProviderFinderQuality implements IProviderFinder {
     } else if (options.size() == 1) {
       return options.get(0);
     } else {
-      return nearest.find(options, person, service, time);
+      // there are a few equally good options, pick one randomly.
+      return options.get(person.randInt(options.size()));
     }
   }
 }
